fix ref levels type

+++ include/ref.hpp
@@ -31,7 +31,7 @@ namespace eosio {
         eosio::name from;
         uint64_t cashback;
         uint8_t level;
-        uint8_t lpercent;
+        uint64_t lpercent;
         uint128_t segments;
         uint64_t primary_key() const {return id;}
